Close the points file after region information has been read
The open_pointfile routine opens the 'points.txt' file as unit 10 to read
the description of the limited-area region, but it never closes this file.
Since no futher reads of the open file are needed, it's best to close
the 'points.txt' file at the end of the open_pointfile routine.